requirements.txt
# Backend dependencies
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.4.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9

# Frontend dependencies
streamlit==1.28.1
requests==2.31.0
pandas==2.1.3
